"Bootstrap 5 Migration - Marked js file 'app_manager/js/forms/case_config_ui.js' as complete and un-split files."
