+++ corehq/apps/app_manager/static/app_manager/js/forms/case_config_ui.js
